Update ge-jasco-zwave-plus-dimmer-switch.groovy
Still problems with the delay at turn on and off events.  Switch ramps as it turns on and off, based on Zwave Steps and Delay.  If delay is too short, switch will send excess level commands as it ramps.  If delay is too long, times out and light on/off status does not update properly.